Outsourced routes/api/articles.js logic to a controller under controllers/articles.js
